Correction des types de données uint64 et int64 pour utiliser des types longs au lieu de longs longs dans types.h

+++ targetDetection/types.h
@@ -10,8 +10,8 @@ typedef signed short int16;
 typedef unsigned int uint32;
 typedef signed int int32;
 
-typedef unsigned long long uint64;
-typedef signed long long int64;
+typedef unsigned long uint64;
+typedef signed long int64;
 
 #ifdef __64_BITS__
 	typedef uint64 SizeT;
